Update troubleshooting.md
I  am having trouble with the Guacamole app and have no idea where the logs to that app might be.  I have looked in /var/log/guac.d and it is empty.  Perhaps others might find this information useful as well.  Many thanks, and sorry if this is not a good suggestion.